BuildSystem_Package/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.6.1
requests>=2.31.0
httpx>=0.26.0
jinja2>=3.1.3

# Model SDKs
anthropic==0.8.1      # Claude API

# Web tools
beautifulsoup4>=4.12.3
tavily-python>=0.3.0

# Testing
pytest==8.0.1
pytest-asyncio>=0.23.5  # For async tests
pytest-cov>=4.1.0      # For coverage reports
httpx>=0.26.0         # For async HTTP
pytest-mock>=3.12.0   # For mocking

# Development and utilities
black==24.2.0
pylint==3.0.3
psutil>=5.9.8
tiktoken>=0.5.2  # Token counting
